+++ test/rails_app/app/misc/rodauth_admin.rb
@@ -1,6 +1,6 @@
 class RodauthAdmin < Rodauth::Rails::Auth
   configure do
-    enable :login
+    enable :login, :webauthn_autofill
     prefix "/admin"
     rails_controller { Admin::RodauthController }
   end